task1/src/movies/Borrowing.java
package movies;

import java.util.Enumeration;
import java.util.Vector;

public class Borrowing {
    private Customer _customer;
    private Vector <Rental> _rentals = new  Vector <> ();
    
    public Borrowing  (Customer customer){
        this._customer = customer;
    }
    
    public Borrowing  (Customer customer, Vector<Rental> rentals){
        this._customer = customer;
        this._rentals = rentals;
    }
    
    public void addRental(Rental arg) {
        _rentals.addElement(arg);
    }
    
    public Vector<Rental> getRentals(){
        return _rentals;
    }
    
    public String statement() {
        double totalAmount = 0;
        int frequentRenterPoints = 0;
        Enumeration<Rental> rentals = _rentals.elements();
        String result = "Rental Record for " + _customer.getName() + "\n";
        while (rentals.hasMoreElements()) {
            double thisAmount = 0;
            Rental each = rentals.nextElement();
          //determine amounts for each line
            thisAmount = typeMovi(thisAmount, each);
            // add frequent renter points
            frequentRenterPoints++;
            // add bonus for a two day new release rental
            if ((each.getMovie().getPriceCode() == Movie.NEW_RELEASE)
                    &&
                    each.getDaysRented() > 1) frequentRenterPoints++;
            //show figures for this rental
            result += "\t" + each.getMovie().getTitle() + "\t" +
                    String.valueOf(thisAmount) + "\n";
            totalAmount += thisAmount;
        }
        //add footer lines
        result += "Amount owed is " + String.valueOf(totalAmount) +
                "\n";
        result += "You earned " + String.valueOf(frequentRenterPoints)
                +
                " frequent renter points";
        return result;
    }

    private double typeMovi(double thisAmount, Rental each) {
        //determine amounts for each line
        switch (each.getMovie().getPriceCode()) {
            case Movie.REGULAR:
                thisAmount += 2;
                if (each.getDaysRented() > 2)
                    thisAmount += (each.getDaysRented() - 2) * 1.5;
                break;
            case Movie.NEW_RELEASE:
                thisAmount += each.getDaysRented() * 3;
                break;
            case Movie.CHILDRENS:
                thisAmount += 1.5;
                if (each.getDaysRented() > 3)
                    thisAmount += (each.getDaysRented() - 3) * 1.5;
                break;
        }
        return thisAmount;
    }

}
